+++ dcpu-hardware-tester.h
@@ -7,8 +7,8 @@
 class TesterDevice : public Hardware {
 public:
     TesterDevice();
-    uint8_t update(DCPU& cpu, Memory& mem) override;
-    uint8_t interrupt(DCPU& cpu, Memory& mem) override;
+    uint32_t update(DCPU& cpu, Memory& mem) override;
+    uint32_t interrupt(DCPU& cpu, Memory& mem) override;
 
 private:
     uint16_t m_lastkey = 0;
